Another explicit cast int for a verbose arg...
git-svn-id: http://svn.mythtv.org/svn/trunk@4345 7dbf422c-18fa-0310-86e9-fd20926502f2

+++ mythtv/libs/libmythtv/avformatdecoder.cpp
@@ -841,7 +841,7 @@ void AvFormatDecoder::HandleGopStart(AVPacket *pkt)
         {
             long long startpos = pkt->startpos;
             VERBOSE(VB_PLAYBACK, QString("positionMap[ %1 ] == %2.").arg(prevgoppos / keyframedist)
-                                         .arg(startpos));
+                                         .arg((int)startpos));
 
             // Grow positionMap vector several entries at a time
             if (m_positionMap.capacity() == m_positionMap.size())
